+++ History.md
@@ -3,6 +3,7 @@ Release date: unreleased
 
 * Dropped support for Ruby 2.7, 3.0+ is now required
 * Dropped support for Selenium < 4.8
+* Support `:data` system filter option
 
 # Version 3.39.2
 Release date: 2023-06-10